test(architecture): move frontend dashboard shell UI batch
