+++ docs/implementation/test-arch-58-frontend-dashboard-shell-ui.md
@@ -1,0 +1,7 @@
+# TEST-ARCH-58 - Frontend dashboard shell UI batch
+
+Moved ten frontend dashboard shell/layout UI tests from root test/ into test/unit/ui/dashboard/.
+
+Updated only test-relative helper imports after relocation.
+
+Guardrails: no runtime/product/API/auth/DB/schema/migration/dependency/lockfile/CI changes.
