+++ test/unit/ui/dashboard/frontend-dashboard-horizontal-nav.test.ts
